requirements.txt
pillow>=11.0.0
numpy>=2.1.2
